complete iterator-factory
