+++ src/lib/iteration/iterable-array.js
@@ -1,5 +1,5 @@
 import Iterable from './iterable';
-import Iterator from './iterator';
+import ArrayIterator from './iterator-array';
 import iteratorSymbol from './iterator-symbol';
 import mixin from '../utils/mixin';
 import extend from '../utils/extend';
@@ -26,33 +26,3 @@ mixin(ArrayIterable.prototype, {
     }
 });
 
-
-/**
-* Supports an iteration over an Array or Array-Like object.
-* @param {Array|String|Array-like} arr An array or array-like object.
-*/
-export function ArrayIterator(arr) {
-    var index = -1,
-        length = arr.length;
-
-    Iterator.call(this, function () {
-        if (++index < length) {
-            return {
-                value: arr[index],
-                done: false
-            };
-        }
-
-        return {
-            done: true
-        };
-    });
-}
-
-extend(ArrayIterator, Iterator);
-
-mixin(ArrayIterator.prototype, {
-    toString: function () {
-        return '[Array Iterator]';
-    }
-});
