+++ src/lib/iteration/iterable.js
@@ -1,9 +1,10 @@
+import iterator from './iterator-factory';
 import iteratorSymbol from './iterator-symbol';
 import mixin from '../utils/mixin';
 
 /**
 * Defines abstract Iterable class.
-* @param {Object} source An Iterable object.
+* @param {Iterable|Array|String|Function|Function*|Object} source An Iterable object.
 */
 export default function Iterable(source) {
     if (source != null) {
@@ -12,7 +13,7 @@ export default function Iterable(source) {
 }
 
 Iterable.prototype[iteratorSymbol] = function () {
-    return this._source[Symbol.iterator]();
+    return iterator(this._source);
 };
 
 mixin(Iterable.prototype, {
