+++ src/lib/iteration/iterator-factory.js
@@ -1,9 +1,45 @@
-import iterable from './iterable-factory';
+import isObject from '../utils/is-object';
+import isFunction from '../utils/is-function';
+import isArrayLike from '../utils/is-array-like';
+import ArrayIterator from './iterator-array';
+import ObjectIterator from './iterator-object';
+import EmptyIterator from './iterator-empty';
 
 /**
 * Creates an iterator object
 * @param {Object} obj An object to create iterator from.
 */
 export default function iterator(obj) {
-    return iterable(obj)[Symbol.iterator]();
+    // empty iteration
+    if (obj === null || obj === undefined) {
+        return new EmptyIterator();
+    }
+
+
+    // iterable/generator function
+    else if (isFunction(obj)) {
+        return obj();
+    }
+
+    // iterable: Array, String, Map, Set, NodeList, Arguments, Iterable objects
+    else if (isFunction(obj[Symbol.iterator])) {
+        return obj[Symbol.iterator]();
+    }
+
+
+    // array-like objects
+    else if (isArrayLike(obj)) {
+        return new ArrayIterator(obj);
+    }
+
+
+    // Object.entries iterator
+    else if (isObject(obj)) {
+        return new ObjectIterator(obj);
+    }
+
+    // simple iterator over non-objects
+    else {
+        return new ArrayIterator([obj]);
+    }
 }
