+++ src/lib/iteration/iterator-object.js
@@ -1,0 +1,37 @@
+import Iterator from './iterator';
+import mixin from '../utils/mixin';
+import extend from '../utils/extend';
+
+/**
+* Supports an iteration over Object properties.
+* @param {Object} obj An object instance.
+*/
+export default function ObjectIterator(obj) {
+    var index = -1,
+        keys = Object.keys(obj),
+        length = keys.length;
+
+    // [key, value] iterator
+    Iterator.call(this, function () {
+        if (++index < length) {
+            return {
+                value: [
+                    keys[index],
+                    obj[keys[index]]
+                ],
+                done: false
+            };
+        }
+        return {
+            done: true
+        };
+    });
+}
+
+extend(ObjectIterator, Iterator);
+
+mixin(ObjectIterator.prototype, {
+    toString: function () {
+        return '[Object Iterator]';
+    }
+});
